treewide: Various cleanups
- Place shell commands for modules in separate files. This makes it
  easier to distribute the shell commands to the correct modules.
  Also removes the need to have a separate shell module.

- Remove config and error channels

- Update tests and references in modules to reflect the changes

- Update tests to use new name and shell commands
