+++ app/CMakeLists.txt
@@ -10,7 +10,7 @@ find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
 project("Asset Tracker Template")
 
 # Include files that are common for all modules
-add_subdirectory(src/common)
+target_include_directories(app PRIVATE src/common)
 
 # Add main application source
 target_sources(app PRIVATE src/main.c)
@@ -26,7 +26,6 @@ add_subdirectory(src/modules/button)
 add_subdirectory_ifdef(CONFIG_APP_POWER src/modules/power)
 add_subdirectory_ifdef(CONFIG_APP_ENVIRONMENTAL src/modules/environmental)
 add_subdirectory_ifdef(CONFIG_APP_LED src/modules/led)
-add_subdirectory_ifdef(CONFIG_APP_SHELL src/modules/shell)
 
 if (CONFIG_NRF_CLOUD_COAP_SEC_TAG GREATER_EQUAL 2147483648 AND CONFIG_NRF_CLOUD_COAP_SEC_TAG LESS_EQUAL 2147483667)
 	message(WARNING "CONFIG_NRF_CLOUD_COAP_SEC_TAG is set to a developer security tag. "
