+++ app/prj.conf
@@ -200,7 +200,6 @@ CONFIG_APP_FOTA_LOG_LEVEL_DBG=y
 CONFIG_APP_LOCATION_LOG_LEVEL_DBG=y
 CONFIG_APP_CLOUD_LOG_LEVEL_DBG=y
 CONFIG_APP_NETWORK_LOG_LEVEL_DBG=y
-CONFIG_APP_SHELL_LOG_LEVEL_DBG=y
 CONFIG_APP_LOG_LEVEL_DBG=y
 
 # Task Watchdog
