--- app/src/common/CMakeLists.txt
@@ -1,9 +1,0 @@
-#
-# Copyright (c) 2024 Nordic Semiconductor ASA
-#
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-#
-
-target_include_directories(app PRIVATE .)
-
-target_sources(app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/message_channel.c)
