--- app/src/common/message_channel.c
@@ -1,27 +1,0 @@
-/*
- * Copyright (c) 2024 Nordic Semiconductor ASA
- *
- * SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
- */
-
-#include <zephyr/kernel.h>
-#include <zephyr/logging/log.h>
-#include <zephyr/zbus/zbus.h>
-
-#include "message_channel.h"
-
-ZBUS_CHAN_DEFINE(ERROR_CHAN,
-		 enum error_type,
-		 NULL,
-		 NULL,
-		 ZBUS_OBSERVERS_EMPTY,
-		 ZBUS_MSG_INIT(0)
-);
-
-ZBUS_CHAN_DEFINE(CONFIG_CHAN,
-		 struct configuration,
-		 NULL,
-		 NULL,
-		 ZBUS_OBSERVERS_EMPTY,
-		 ZBUS_MSG_INIT(0)
-);
