+++ app/src/modules/button/CMakeLists.txt
@@ -5,4 +5,5 @@
 #
 
 target_sources(app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/button.c)
+target_sources_ifdef(CONFIG_APP_BUTTON_SHELL app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/button_shell.c)
 target_include_directories(app PRIVATE .)
