+++ app/src/modules/cloud/CMakeLists.txt
@@ -5,4 +5,5 @@
 #
 
 target_sources(app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/cloud_module.c)
+target_sources_ifdef(CONFIG_APP_CLOUD_SHELL app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/cloud_module_shell.c)
 target_include_directories(app PRIVATE .)
