+++ app/src/modules/network/CMakeLists.txt
@@ -5,4 +5,5 @@
 #
 
 target_sources(app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/network.c)
+target_sources_ifdef(CONFIG_APP_NETWORK_SHELL app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/network_shell.c)
 target_include_directories(app PRIVATE .)
