--- app/src/modules/shell/CMakeLists.txt
@@ -1,7 +1,0 @@
-#
-# Copyright (c) 2024 Nordic Semiconductor
-#
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-#
-
-target_sources(app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/shell.c)
