--- app/src/modules/shell/Kconfig.shell
@@ -1,22 +1,0 @@
-#
-# Copyright (c) 2024 Nordic Semiconductor
-#
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-#
-
-menuconfig APP_SHELL
-	bool "Zbus Shell"
-	depends on SHELL
-	depends on !AT_HOST_LIRBARY
-	imply AT_SHELL
-	default y
-	help
-	  Shell that adds usefaul commands to control certain aspects of the application.
-
-if APP_SHELL
-
-module = APP_SHELL
-module-str = SHELL
-source "subsys/logging/Kconfig.template.log_config"
-
-endif # APP_SHELL
