+++ tests/module/fota/src/fota_module_test.c
@@ -4,14 +4,13 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
  */
 #include <unity.h>
-
 #include <zephyr/fff.h>
 #include <zephyr/zbus/zbus.h>
 #include <zephyr/task_wdt/task_wdt.h>
 #include <zephyr/logging/log.h>
 #include <net/nrf_cloud_fota_poll.h>
 
-#include "message_channel.h"
+#include "app_common.h"
 #include "fota.h"
 
 DEFINE_FFF_GLOBALS;
