+++ tests/module/main/CMakeLists.txt
@@ -18,7 +18,6 @@ target_sources(app
 	src/main.c
 	src/checks.c
 	${ASSET_TRACKER_TEMPLATE_DIR}/app/src/main.c
-	${ASSET_TRACKER_TEMPLATE_DIR}/app/src/common/message_channel.c
 )
 
 target_include_directories(app PRIVATE src)
@@ -48,7 +47,8 @@ target_compile_definitions(app PRIVATE
 	-DCONFIG_APP_LOG_LEVEL=4
 	-DCONFIG_APP_MSG_PROCESSING_TIMEOUT_SECONDS=120
 	-DCONFIG_APP_WATCHDOG_TIMEOUT_SECONDS=180
-	-DCONFIG_APP_PAYLOAD_CHANNEL_BUFFER_MAX_SIZE=100
+	-DCONFIG_APP_CLOUD_PAYLOAD_BUFFER_MAX_SIZE=100
+	-DCONFIG_APP_CLOUD_SHADOW_RESPONSE_BUFFER_MAX_SIZE=512
 	-DCONFIG_APP_MODULE_TRIGGER_TIMEOUT_SECONDS=600
 	-DCONFIG_APP_CLOUD_LOG_LEVEL=0
 	-DCONFIG_APP_CLOUD_THREAD_STACK_SIZE=2048
