+++ tests/module/network/CMakeLists.txt
@@ -15,7 +15,6 @@ target_sources(app
   PRIVATE
   src/main.c
   ../../../app/src/modules/network/network.c
-  ../../../app/src/common/message_channel.c
 )
 
 zephyr_include_directories(${ZEPHYR_BASE}/include/zephyr/)
@@ -27,7 +26,7 @@ zephyr_include_directories(../../../app/src/modules/network)
 target_link_options(app PRIVATE --whole-archive)
 # Options that cannot be passed through Kconfig fragments
 target_compile_definitions(app PRIVATE
-	-DCONFIG_APP_PAYLOAD_CHANNEL_BUFFER_MAX_SIZE=100
+	-DCONFIG_APP_CLOUD_PAYLOAD_BUFFER_MAX_SIZE=100
 	-DCONFIG_APP_NETWORK_LOG_LEVEL=4
 	-DCONFIG_APP_NETWORK_THREAD_STACK_SIZE=1024
 	-DCONFIG_APP_NETWORK_MESSAGE_QUEUE_SIZE=5
