+++ tests/on_target/README.md
@@ -47,7 +47,7 @@ export NRFCLOUD_API_KEY=<your_nrfcloud_api_key>
 Run desired tests, example commands
 ```shell
 pytest -s -v -m "not slow" tests
-pytest -s -v -m "not slow" tests/test_functional/test_uart_output.py
+pytest -s -v -m "not slow" tests/test_functional/test_network_reconnect.py
 pytest -s -v -m "not slow" tests/test_functional/test_sampling.py
 pytest -s -v -m "slow" tests/test_functional/test_fota.py::test_full_mfw_fota
 ```
