+++ tests/on_target/tests/test_functional/test_uart_output.py
@@ -3,10 +3,7 @@
 # SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
 ##########################################################################################
 
-import pytest
-import time
 import os
-import re
 from utils.flash_tools import flash_device, reset_device
 import sys
 sys.path.append(os.getcwd())
@@ -14,7 +11,7 @@ from utils.logger import get_logger
 
 logger = get_logger()
 
-def test_uart_output(dut_board, hex_file):
+def test_network_reconnect(dut_board, hex_file):
     '''
     Test that the device is operating normally by checking UART output
     '''
@@ -24,14 +21,10 @@ def test_uart_output(dut_board, hex_file):
         "Network connectivity established",
         "Connected to Cloud",
     ]
-    patterns_button_press = [
-        "trigger: frequent_poll_run: Button 1 pressed in frequent poll state, restarting duration timer",
-        "trigger_poll_work_fn: Sending shadow/fota poll trigger"
-    ]
-    patterns_lte_offline = [
+    patterns_network_disconnected = [
         "network: Network connectivity lost",
     ]
-    patterns_lte_normal = [
+    patterns_network_connected = [
         "network: Network connectivity established",
     ]
 
@@ -40,17 +33,12 @@ def test_uart_output(dut_board, hex_file):
     reset_device()
     dut_board.uart.wait_for_str(patterns_boot, timeout=120)
 
-    # # Button press
-    # dut_board.uart.flush()
-    # dut_board.uart.write("zbus button_press\r\n")
-    # dut_board.uart.wait_for_str(patterns_button_press, timeout=120)
+    # LTE disconnect
+    dut_board.uart.flush()
+    dut_board.uart.write("att_network disconnect\r\n")
+    dut_board.uart.wait_for_str(patterns_network_disconnected, timeout=20)
 
-    # # LTE disconnect
-    # dut_board.uart.flush()
-    # dut_board.uart.write("lte offline\r\n")
-    # dut_board.uart.wait_for_str(patterns_lte_offline, timeout=20)
-
-    # # LTE reconnect
-    # dut_board.uart.flush()
-    # dut_board.uart.write("lte normal\r\n")
-    # dut_board.uart.wait_for_str(patterns_lte_normal, timeout=120)
+    # LTE reconnect
+    dut_board.uart.flush()
+    dut_board.uart.write("att_network connect\r\n")
+    dut_board.uart.wait_for_str(patterns_network_connected, timeout=120)
